requirements.txt
flask
pyodbc==5.0.1
azure-cognitiveservices-speech==1.36.0
